Rename variables and views, add comments describing logic
