+++ ios-swiftui/realm-template-apps-swiftui.xcodeproj/xcuserdata/dachary.carey.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -49,7 +49,7 @@
 		<key>realm-template-apps-swiftui.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
 		</dict>
 	</dict>
 </dict>
