+++ ios-swiftui/realm-template-apps-swiftui/Views/ContentView.swift
@@ -18,13 +18,3 @@ struct ContentView: View {
             .padding()
     }
 }
-
-struct ErrorView: View {
-    @State var error: Error
-        
-    var body: some View {
-        VStack {
-            Text("Error: \(error.localizedDescription)")
-        }
-    }
-}
