+++ ios-swiftui/realm-template-apps-swiftui/Views/ErrorView.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct ErrorView: View {
+    @State var error: Error
+        
+    var body: some View {
+        VStack {
+            Text("Error: \(error.localizedDescription)")
+        }
+    }
+}
