+++ ios-swiftui/realm-template-apps-swiftui/Views/LogoutButton.swift
@@ -4,19 +4,36 @@ import RealmSwift
 /// Logout from the synchronized realm. Returns the user to the login/sign up screen.
 struct LogoutButton: View {
     @State var isLoggingOut = false
+    @State var error: Error?
     
     var body: some View {
+        if isLoggingOut {
+            ProgressView()
+        }
+        if let error = error {
+            Text("Error: \(error.localizedDescription)")
+        }
         Button("Log Out") {
             guard let user = realmApp.currentUser else {
                 return
             }
             isLoggingOut = true
-            user.logOut() { error in
+            Task.init {
+                await logout(user: user)
                 isLoggingOut = false
                 // Other views are observing the app and will detect
                 // that the currentUser has changed. Nothing more to do here.
-                print("Logged out")
             }
         }.disabled(realmApp.currentUser == nil || isLoggingOut)
+    }
+    
+    func logout(user: User) async {
+        do {
+            try await user.logOut()
+            print("Successfully logged user out")
+        } catch {
+            print("Failed to log user out: \(error.localizedDescription)")
+            self.error = error
+        }
     }
 }
