+++ ios-swiftui/realm-template-apps-swiftui/realm_template_apps_swiftuiApp.swift
@@ -1,5 +1,3 @@
-//let realmAppId = Bundle.main.object(forInfoDictionaryKey: "REALM_APP_ID") as? String
-
 import SwiftUI
 import RealmSwift
